add room1.py modif main to do a launcher
